Se fixean bugs, y se realiza la persistencia en la base de datos, se cambia el current date ahora es administrado por la base de datos

+++ TexasHoldemPoker/src/texasHoldemPoker/Persistence/Sql/Dao/ISalaryHistoryDAO.java
@@ -6,5 +6,5 @@ import texasHoldemPoker.Model.SalaryHistory;
 
 public interface ISalaryHistoryDAO {
 	ArrayList<SalaryHistory> GetByPlayerId(int playerId);
-	void AddBalance(double amount, double balance, int playerId);
+	void AddBalance(int amount, int balance, int playerId);
 }
